Close COPY TO/FROM PROGRAM descriptors in case of error (#945)
This is a port of the 75981eb. Original commit message:

Commit 110b825 added an ifThrow argument to the close_program_pipes function.
That commit contained a call to close_program_pipes(cstate, false) on errors.
But then this function call was lost. And now, in case of errors, open pipes are
not closed until the end of the session.

This patch adds a callback for closing handles on errors, which is called before
deleting the memory context of the copy state.

Co-authored-by: Georgy Shelkovy <g.shelkovy@arenadata.io>